chore: improve lcp
